% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_palette.R
\name{create_palette}
\alias{create_palette}
\title{make a palette from your image}
\usage{
create_palette(image_path = NA, number_of_colors = 20)
}
\arguments{
\item{image_path}{relative path to the custom image, including .jpeg extension}

\item{number_of_colors}{number of different colors desired for the resulting palette}
}
\description{
processes a custom jpeg image producing a palett with a user-defined number of colours
}
\examples{
create_palette("data/nascita_venere.jpg",number_of_colors = 20)
}
\author{
Andrea Cirillo
}
