% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_maker.R
\name{palette_maker}
\alias{palette_maker}
\title{make a palette from your image}
\usage{
palette_maker(image_path = NA, number_of_colors = 20)
}
\arguments{
\item{image_path}{relative path to the custom image, including .jpeg extension}

\item{number_of_colors}{number of different colors desired for the resulting palette}
}
\description{
processes a custom jpeg image producing a palett with a user-defined number of colours
}
\examples{
palette_maker("data/nascita_venere.jpg",number_of_colors = 20)
}
\author{
Andrea Cirillo
}
