% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_maker.R
\name{palette_maker_cs}
\alias{palette_maker_cs}
\title{Make a palette from your image in a particular color space}
\usage{
palette_maker_cs(image_path = NA, number_of_colors = 9, colspace = "RGB",
  iter.max = 50, nstart = 5, plot_colors = TRUE, seed = NULL,
  use_hex = FALSE)
}
\arguments{
\item{image_path}{Relative path to the custom image, including .jpeg
extension}

\item{number_of_colors}{Number of different colors desired for the resulting
palette}

\item{colspace}{A valid color class from the `colorspace` package. Valid
choices are 'RGB', 'LAB', 'LUV', 'XYZ', 'HLS', 'HSV', 'polarLUV',
'polarLAB', 'sRGB'.}

\item{iter.max}{Number of local iterations of kmeans}

\item{nstart}{Number of global re-runs of kmeans}

\item{plot_colors}{Whether to display a rough plot of the selected colors
using the `scales` package.}

\item{seed}{Set a seed for reproducibility.}

\item{use_hex}{Whether to use the `colorspace::hex()` function or
`grDevices::rgb()` for converting to a hex color string when returning the
result. `hex()` seems to return lighter, pastel-like colors.}
}
\description{
Processes a custom jpeg image producing a palett with a user-defined number
of colours. This will allow any color space available in the `colorspace`
package. There may be conversion issues if the selected color space does not
have a valid representation of the all RGB values in the picture.
}
\examples{
  palette_maker_cs(
    system.file('extdata', "nascita_venere.jpg", package="paletter"),
    number_of_colors = 20, colspace = "LAB", iter.max = 100)
}
\author{
Victor Lei (based on Andrea Cirillo's original code)
}
