% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_maker.R
\name{run_test_paletter}
\alias{run_test_paletter}
\title{Use a number of different methods to generate color palettes}
\usage{
run_test_paletter(img_path, out_prefix, test_col_spaces = c("RGB", "LAB",
  "HLS"), num_colors = 9, out_w = 500, out_h = 500, nstart = 1,
  iter.max = 100)
}
\arguments{
\item{img_path}{The path to your JPEG image file}

\item{out_prefix}{The filename output prefix to which the color space name is
appended (e.g. 'myimage' results in 'myimage_RGB.png', 'myimage_LAB.png'
etc.)}

\item{test_col_spaces}{Which color spaces to test. All valid color spaces in the `colorspace` package are supported. See \code{\link{palette_maker_cs}}.}

\item{num_colors}{Number of colors to produce in your palette}

\item{out_w}{Output image width in pixels}

\item{out_h}{Output image height in pixels}

\item{nstart}{Number of global re-runs of kmeans}

\item{iter.max}{Number of local iterations of kmeans}
}
\description{
Clustering in different color spaces yield somewhat different color palettes.
This helper function automates the process of clustering your image in
different color spaces.
}
\examples{
 run_test_paletter('data/nascita_venere.jpg', 'nascita_venere', num_colors = 9, nstart = 5)
}
